requirements.txt
streamlit==1.29.0
pymongo==4.6.1
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==3.17.4
huggingface_hub==0.15.1
neo4j==5.15.0
pandas==2.1.4
